refactor: Cleaning up and adding docs
